cc-wrapper: -nostdlib doessn't mean not C++
This was introduced in 87607af7a1bf35682f8ad206307ed46e8ead260a, which
made the wrapper aware of `-x <lang>` flags. I am not sure why this new
case was added, especially with the `-l` that looks like a mistake.

FWIW, https://github.com/briansmith/nostdlib is an example of writing C++
with -nosdlib.