alacritty: manually strip before running patchelf
patchelf seems to generate an ELF that "strip" cannot process:

  strip: /tmp/stNGAZvd: not enough room for program headers, try linking with -N
  strip:/tmp/stNGAZvd[.interp]: bad value

(This unfortunately does not cause any errors or even error output in
the build log, since the default strip hook redirects them to /dev/null
and ignores failures.)

The lack of stripping ends up leaking paths to the compiler in the debug
symbols. This more than doubles the transitive closure size of
alacritty.

Impact on closure size: 314.5M -> 131.8M.