aflplusplus: init at 2.64c
This commit adds the aflplusplus package.
The aflplusplus package is a fork of the afl package, which is already
included in the nixpkgs repository. Currently, only source fuzzing mode
is supported, as I'm not too interested in supporting the qemu fuzzing
mode.

If anyone wants to take a stab at it, make sure to make a separate
derivation for the qemu build and, ideally, make it so that the qemu
mode/build can be disabled. This prevents issues where aflplusplus
can't be used anymore even if you don't use the qemu mode.