nixos/datadog-agent: Fix restartTriggers
Fixes #85800

1d61efb7f177f7b70c467ab4940fde0a3481d4dc accidentially changed the
restartTriggers of `datadog-agent.service` to point to the attribute
name (in this case, a location relative to `/etc`), instead of the
location of the config files in the nix store.

This caused datadog to not get restarted on activation of new
config, if the file name hasn't changed.

Fix this, by pointing this back to the location in the nix store.

(cherry picked from commit f332109ebf67951b412457fee023222c4ed5f561)