rustc: 1.42.0 -> 1.43.0
Kept 1.42 around for Thunderbird.

i686-apple-darwin is no longer supported upstream.  We could still
support building it, for this one release, since we have the binary
for the previous release, (or bootstrap it for future releases from
Rust 1.42,) but since this release is the one that drops support, I
think it makes sense to do it now.  (And probably nobody is using it
anyway.)