ycmd: typescript completion support
Link the already-packaged typescript package so that ycmd will support
completions for javascript and typescript filetypes.

The structure of these installation directories is "documented" here:
<https://github.com/ycm-core/ycmd/blob/master/build.py>.

I don't completely understand why the language support is passed as
optional attrs Is it because folks want to build leaner packages and
pass null or because they want to target particular dependencies? At any
rate I followed the pattern.