dotnet: add initial support for arm64
`jellyfin` appeared unsupported on `aarch64` due to `dotnet` platform
support in nixpkgs, but there are ARM64 downloads of the `dotnet` SDKs
available. This change follows the kind of pattern used in the
`firecracker` packaging to support selective x86_64/arm64 downloads.

With this change I can build `jellyfin` on a Raspberry Pi 4. The other
content hashes have been filled in, and all build successfully, but
they have not been further tested.