coreutils: set ac_cv_func_lchmod=no
Bumping binutils to 2.32 broke coreutils for unknown reasons[1]. Upon
further investigation we found that there was some issue with mknod
inside a chroot[2][3], which setting ac_cv_func_lchmod to "no" _somehow_
fixes.

We apply this fix and hope to never have to debug this again.

Thanks to @flokli and @lnl7 who helped me chase after this and @pbogdan
who found the fix in the fedora build.

[1]: https://github.com/NixOS/nixpkgs/pull/85951
[2]: https://github.com/NixOS/nixpkgs/pull/85951#issuecomment-619601978
[3]: https://bugzilla.redhat.com/show_bug.cgi?id=1811038