i3-gaps: Fix man pages and simplify
The .tar.gz from the github's archive does not contain man pages, only
asciidoc versions of them. Because i3-gaps uses the same build process
as i3, the man pages are not generated and the asciidoc versions are
put in share/man/man1. This annoys mandb:

mandb: warning: …/share/man/man1/i3.man.gz: ignoring bogus filename

This commit changes the downloaded file to use github's release
instead. The resulting .tar.bz2 file is much closer to the one
downloaded for i3 which means the build process can still be the same
and we get proper man pages at the end.

Because of the previous change, the I3_VERSION file is now part of the
downloaded source which means Nix doesn't have to create it anymore.