vscode-extensions.ms-vscode.cpptools: 0.27.0 -> 0.27.1
Extension is now fetched from the Microsoft repository (instead of the
marketplace). It also contains platform-specific binary dependencies. In the
previous version they were fetched from the Microsoft website and additionally
inserted to the final extension.

Package supports Linux, but there is probably a way no make it work on OSX.