nixos/lib/make-ext4-fs: always allow execution of resize2fs
Without the `-f` flag, `resize2fs` checks that the last time where the
filesystem was checked is greater or equal than the last time is mounted.
This currently works due to a bug in `cptofs`
(https://github.com/lkl/linux/issues/488) and it breaks on slower systems
when the last mount time is greater than 1970-01-01 00:00:01.

To fix this, either we allow `fsck` to update the "last checked" timestamp
on the filesystem, which requires to replace the `-n` option with `-y`,
or we force `resize2fs` to skip these checks. Since we're only dealing
with an image file which is checked immediately before resizing, it is
safe to bypass this check as `fsck` will fail if the filesystem is unclean.