apitrace: fix OpenGL library path and Qt GUI
For unknown reasons, Apitrace breaks the runpath of the traced program.
OpenGL programs on nix are looking at /run/opengl-driver/lib to find the
system OpenGL driver.

This change:

- add path /run/opengl-driver{-32} to the RPATH of the wrapper
libraries. This fixs tracing with `apitrace trace`.
- add rpath to `libglnvd` and `libGL` to the `{e}glretrace` binaries.
These libraries are loaded at runtime, but does not appears as NEEDED in
the binaries, hence the need for the explicit rpath addition. This fix
`{e}glretrace`.
- Explicitly add Qt wrapper to `qapitrace`. It fixs the GUI.

Co-authored-by: Florian Klink <flokli@flokli.de>