sooperlooper: Add alsa library build dependency
Currently, sooperlooper is built without alsa support:

    checking alsa/asoundlib.h usability... no
    checking alsa/asoundlib.h presence... no
    checking for alsa/asoundlib.h... no

This causes sooperlooper to have no midi ports, which prevents it from
being controlled by a hardware midi controller.

This change adds alsaLib as a build dependency. With this change the package
is build with support for alsa midi ports.

This is my very first nixpkgs pull request, if anything is suboptimal I will
happily fix it.