nco: Patch hardcoded paths cp and mv
Commit b83ed3efe81525302469a3052982f41393a616e0 in the nco source
repo, hard-coded the paths to cp and mv, so as to point to their FHS
locations, which resulted in ncrename throwing a "file not found"
error when invoked outside of an FHS sandbox.

This nixpkgs commit patches the src/nco_fl_utl.c to instead point the
associated nixpkgs.coreutils locations. The hard-coded paths actually
occur in a number of other locations, but this is enough to quash the
error w/ ncrename (and presumably the other CLI nco tooling, although
this isn't tested).