nixos/tests/gitdaemon: fix spurious test failures due to flaky network
This test is sometimes flaky on hydra as at the time of the `git clone`
the network isn't really configured yet[1]. That problem doesn't seem to
occur locally but if you run it on a machine with high enough load (such
as hydra build machines). Hopefully this will make the test not flaky
anymore.

[1] https://hydra.nixos.org/build/118710378/nixlog/21/raw