treewide: add editorCommand passthru
This adds a new passthru to all text editor derivations, analogous to
the shellPath passthru, containing a string suitable to be used as the
$EDITOR environment variable. Editors that daemonise by default are
passed flags that stop them from doing so.

This covers every single text editor in nixpkgs. IDEs were not given
such treatment, with the rationale that they are intended to be used
for editing entire projects - and often can't be configured to open
single files or disable daemonisation. Similarly, task-specific text
editors (such as Markdown editors) were skipped.