nixos/defaults: init
This introduces a new set of options for configuring the default editor
and pager. users.defaults.editor can be set to any derivation with the
'editorCommand' attribute, which is a string suitable for being used as
the $EDITOR variable. Similarly, the option users.defaults.pager can be
set to any derivation with the 'pagerCommand' attribute, and controls
the $PAGER variable.

Of note: programs invoking $EDITOR and $PAGER expect the respective
program to not exit until the user has finished editing/viewing the
file, which means that any program that daemonises by default (such as
Atom or Sublime Text) needs an additional flag to be passed. Such flags
are specific to the program in question, and are usually documented in
manual pages and --help output.

This also means that the system path no longer implicitly contains nano
and less, which is probably welcome news for anyone with particularly
passionate opinions about text editors ;)