chrome-gnome-shell: Fix missing introspection data
It was crashing with:

    TypeError: gobject `__main__+ChromeGNOMEShell' doesn't support property `application_id'

– that is the constructor of parent class Gio.Application does not recognize the kwarg.

This is typically caused by missing pygobject overrides but I think ones for Gio are built-in into pygobject.

Even weirder, adding just `${gobject-introspection}/lib/girepository-1.0` seems to fix the issue so it might be missing the whole typelib.
But then why does not it fail when importing it?

	from gi.repository import GLib, Gio

For now, I am adding the Gio typelib which should been done from the start but more debugging should be done since weirdness like this can bring more bugs.

Fixes: https://github.com/NixOS/nixpkgs/issues/87740