sublime3{,-dev}: fix updateScript
The HTML page changed and the update script was no longer finding the latest version.
I changed it to query for the latest version using dedicated URL:

https://forum.sublimetext.com/t/fixed-download-url-for-latest-version/23763/7

Also terminated the update sooner if we detect no update available.
update-source-version handles that normally but we are cleaning the version
so that we can update both platforms at the same time.