flashtool: remove
This package has been broken since 2017, essentially untouched since
2014, and has no maintainers.

The last upstream release was in 2016; the download link for that
release is dead. Although i was able to download it via BitTorrent,
that sounds hard to integrate into NixOS. Additionally, the latest
phone the program works with was released in 2013, so I suspect the
demand is fairly minimal.

Given the amount of TODOs in the derivation and its long-broken state, I
conclude that any update of this package would likely be as much effort
as packaging it from scratch, and since nobody as stepped up to do so in
several years, it can be safely removed.