Make sure glib and gtk3 are both in nativeBuildInputs
I don't know why this didn't break before, but glib and (or) gtk3 need
to be in nativeBuildInputs for glib-compile-schemas to be available. But
they also need to be in buildInputs to be available at link time.

I think worldofpeace mentioned this to me, but that it got lost in all
of the other fixes I needed to make when I first submitted the
derivation.