amarok: 2.9.0-20190824 -> amarok-unstable 2020-06-12
This commit bumps amarok to the most recent commit in master and adds liblastfm
as a dependency to make use of the last.fm integration that has been re-enabled
in upstream master (it was disabled for a while due to breakage on Qt5).

I also updated the package name and version to match [the stipulations in the
Nixpxgs manual](https://nixos.org/nixpkgs/manual/#sec-package-naming).